refactor: 클래스 네이밍 변경
- 의미를 가지는 네이밍을 기준으로 변
