--- src/main/java/racing/RacingGameImpl.java
@@ -1,12 +1,0 @@
-package racing;
-
-/**
- * @author : 김윤호
- * @version : 1.0
- * @date : 2019-10-28 10:28
- */
-public interface RacingGameImpl {
-    public void start();
-    public int getRandomIntValue();
-    public int getParticipantCarCount();
-}
